Add files via upload
The aim is to classify iris flowers among three species (Setosa, Versicolor, or Virginica) from sepals' and petals' length and width measurements.
The iris data set contains fifty instances of each of the three species.
The central goal is to design a model that makes proper classifications for new flowers. In other words, one which exhibits good generalization.